some fixing of fallout from moving stuff into LibGDX.pm

+++ lib/IndieRunner.pm
@@ -129,7 +129,7 @@ $game_name = 'unknown' unless $game_name;
 my $steam_appid = steam_appid();
 my $child_steamlog;
 if ( $steam_appid ) {
-	say 'Found steam_appid: ' . $steam_appid;
+	say 'Found steam_appid: ' . $steam_appid if $verbose;
 	$child_steamlog = log_steam_time $steam_appid if cli_log_steam_time();
 }
 
@@ -169,7 +169,7 @@ write_file( $merged_out, $logfile ) if $merged_out;
 # XXX: inspect $merged_out
 
 # clean up and exit
-kill 'KILL', $child_steamlog;
+kill 'KILL', $child_steamlog if $child_steamlog;
 exit;
 
 1;
